Making the startup process a little cleaner with config.json
Created a config module that requires a json file with the videos path configured, and made it required for that to be configured for the product to even run. I also made it so that the findVideos function only runs when it needs to since theres no reason for it to run when that file gets required.