windows powershell script to setup nodejs-arch
This is the windows variant of the bash script used to create nodejs-arch docker image for homeflix. It is up to you to get docker setup in windows using linux containers properly.... Im working on making the ability to recursively locate all folders that house videos on a machine and make many volume mounts to point into homeflix video folder in image (which will have subdirectories created reflecting names of video directories discovered from folder search) that will probably be next commit but only for linux because Windows is nothing but headaches and much loss in sanity. ....